requirements.txt

# Core dependencies
langchain>=0.1.0
transformers>=4.30.0
faiss-cpu>=1.7.4
groq>=0.4.0
python-dotenv>=1.0.0
sentence-transformers>=2.3.0
torch>=2.0.0

# API dependencies
fastapi>=0.103.1
uvicorn>=0.23.2
pydantic>=2.3.0

# UI dependencies
gradio>=4.8.0

# Search dependencies
google-search-results>=2.4.2

# Data handling
datasets>=2.14.0